fix flexGrow is float & less than 1

+++ src/config.js
@@ -102,8 +102,8 @@ class Config {
     if(typeof flex === 'number') {
       this.flexGrow = this.flexGrow || flex;
     }
-    this.flexShrink = parseInt(this.flexShrink, 10) || 1;
-    this.flexGrow = parseInt(this.flexGrow, 10) || 0;
+    this.flexShrink = parseFloat(this.flexShrink, 10) || 1;
+    this.flexGrow = parseFloat(this.flexGrow, 10) || 0;
     let flexBaxis = this.flexBaxis;
     if(flexBaxis) {
       const flexDirection = this.node.parent.flexDirection;
